Hải (Bán hàng add sp + thanh toán)

+++ datn/src/main/java/nice/store/datn/entity/SanPhamChiTiet.java
@@ -58,7 +58,7 @@ public class SanPhamChiTiet {
     private String maSpct;
 
     @Column(name = "GIA_BAN")
-    private Double giaBan;
+    private BigDecimal giaBan;
 
     @Column(name = "SO_LUONG")
     private Integer soLuong;
