+++ datn/src/main/java/nice/store/datn/response/SanPhamCTResponse.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -21,7 +22,7 @@ public class SanPhamCTResponse {
     private String tenDeGiay;
     private String tenThuongHieu;
 
-    private Double giaBan;
+    private BigDecimal giaBan;
     private Integer soLuong;
     private String moTa;
     private LocalDateTime ngayTao;
